Middle

+++ public/javascripts/dashboard_teacher.js
@@ -53,34 +53,28 @@ $(function () {
       var userInput = $('#js-modal-delete-input').val()
 
       // check if user input of class name matches the actual class name they are trying to delete
-      if ()
+      if (userInput === class_name) {
+        $.ajax('/api/class/' + class_uuid, {
+          type: 'DELETE'
+        }).then(
+          function () {
+            console.log('deleted class', class_uuid)
+            window.location.reload()
+          }
+        )
+      } else {
+
+      }
     })
 
-    /** 
-     * waaay later 
+    /**
+     * waaay later
      */
-    $.ajax('/api/class/' + id, {
-      type: 'DELETE'
-    }).then(
-      function () {
-        console.log('deleted class', id)
-        window.location.reload()
-      }
-    )
   })
 
   // choose a class to view
   $('#view_class').on('click', function (event) {
     var id = $(this).data('id')
-
-    $.ajax('/api/class/' + id, {
-      type: 'GET'
-    }).then(
-      function () {
-        console.log('')
-        window.location.reload()
-      }
-    )
   })
   // delete a class
   $('#delete_class').on('click', function (event) {
